+++ relay-event-normalization/src/transactions/mod.rs
@@ -1,5 +1,5 @@
-mod processor;
 mod rules;
+mod utils;
 
-pub use processor::*;
 pub use rules::*;
+pub use utils::*;
